Probabilistic/l05-monday19-BlonskiP/requirements.txt
jupyter==1.0.0
matplotlib==3.2.0
pyro-ppl==1.3.0
torch>=1.4.0
scikit-learn==0.22.2.post1
seaborn==0.10.0
